+++ FE/src/Components/KYH/admin/ProductList/Table.tsx
@@ -3,7 +3,7 @@ import TableTitle from './TableTitle';
 
 export default function Table() {
   return (
-    <div className="bg-gray-100 border border-gray-400 rounded-t-lg w-250 h-128">
+    <div className="bg-gray-100 border border-gray-400 rounded-t-lg w-300 h-128">
       <TableTitle />
       <ul>
         <TableContents />
