+++ FE/src/Components/KYH/admin/ProductList/TableContents.tsx
@@ -1,13 +1,14 @@
 export default function TableContents() {
   return (
-    <li className="flex items-center w-full h-12 border-b border-b-gray-400">
-      <input type="checkbox" className="w-6 h-6 ml-10" />
-      <p className="ml-12 mr-10 text-gray-700 truncate w-104">
+    <li className="flex items-center justify-between h-12 px-6 border-b border-b-gray-400">
+      <input type="checkbox" className="w-5 h-5" />
+      <span>1</span>
+      <p className="text-gray-700 truncate w-120">
         알파스캔 AOC Q32V3S QHD IPS 75 시력보호 무결점 SRGB 100% 전문가용 모니터
       </p>
-      <span className="w-16 text-center mr-14">150</span>
-      <span className="w-12 text-center mr-14">50</span>
-      <span className="w-12 -ml-1 text-center mr-11">100</span>
+      <span className="w-16 text-center">150</span>
+      <span className="w-12 text-center">50</span>
+      <span className="w-12 text-center">100</span>
       <span className="w-24 text-center">100,000,000</span>
     </li>
   );
